+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 0.16.0
 
 - Update Camel Quarkus Catalog from 3.6.0 to 3.7.0
+- Provide entry in `File > New file...` menu to create a Camel file
 
 ## 0.15.0